requirements.txt
django==5.1.1
pillow==10.4.0
djangorestframework==3.15.2
python-decouple==3.8
psycopg2-binary==2.9.10
dj-database-url==2.2.0
django-filter==24.3
django-treebeard==4.7.1
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
django-treebeard==4.7.1
python-ulid==2.7.0
pytz==2024.2
boto3==1.35.34
django-storages==1.14.4
celery==5.4.0
django-celery-beat==2.7.0
flower==2.0.1
django_celery_results==2.5.1
urllib3==2.2.3
botocore==1.35.34
colorlog==6.8.2
django_redis==5.4.0
msgpack==1.1.0
pyzstd==0.16.1
requests==2.32.3
django-cors-headers==4.4.0
gunicorn==23.0.0
whitenoise==6.7.0
colorlog==6.8.2
dj-logging==1.2.1
django-debug-toolbar==4.4.6
django-jalali==7.1.0
hiredis==3.0.0
redis==5.1.1
pyzstd==0.16.1
